Extract CLI commands into a separate module
